Extract event edit drawer into a separate component and add tests
